// WADNR.Web/src/app/shared/generated/model/gdb-import-basics-upsert-request.ts
export class GdbImportBasicsUpsertRequest { 
    ProjectTypeDefaultName?: string | null;
    TreatmentTypeDefaultName?: string | null;
    ImportIsFlattened?: boolean | null;
    AdjustProjectTypeBasedOnTreatmentTypes?: boolean;
    ProjectStageDefaultID?: number;
    DataDeriveProjectStage?: boolean;
    DefaultLeadImplementerOrganizationID?: number;
    RelationshipTypeForDefaultOrganizationID?: number;
    ImportAsDetailedLocationInsteadOfTreatments?: boolean;
    ImportAsDetailedLocationInAdditionToTreatments?: boolean;
    ProjectDescriptionDefaultText?: string | null;
    ApplyStartDateToProject?: boolean;
    ApplyCompletedDateToProject?: boolean;
    ApplyStartDateToTreatments?: boolean;
    ApplyEndDateToTreatments?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}